LayoutTests/http/tests/xmlhttprequest/xml-encoding-expected.txt

Test XMLHttpRequest and document.xmlEncoding interaction.

The request body should contain the word "Проверка".

responseXML.xmlEncoding: windows-1251
responseXML.xmlVersion: 1.0
responseXML.xmlStandalone: true
Request body: <doc>Проверка</doc>
